# api/openapi.yaml
openapi: "3.1.0"
info:
  title: You2Midi API
  description: YouTube piano video to MIDI transcription service
  version: "0.1.0"

paths:
  /jobs:
    get:
      operationId: listJobs
      summary: List jobs (newest first)
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          description: Maximum number of jobs to return.
        - name: before
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Return only jobs created before this timestamp (RFC3339).
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      summary: Submit a new transcription job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          $ref: "#/components/responses/BadRequest"

  /jobs/{id}:
    get:
      operationId: getJob
      summary: Get a job by ID
      parameters:
        - $ref: "#/components/parameters/JobID"
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          $ref: "#/components/responses/NotFound"

  /jobs/{id}/cancel:
    post:
      operationId: cancelJob
      summary: Cancel a running or queued job
      parameters:
        - $ref: "#/components/parameters/JobID"
      responses:
        "200":
          description: Job cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  /jobs/{id}/progress:
    get:
      operationId: getJobProgress
      summary: Get job progress snapshot for polling
      parameters:
        - $ref: "#/components/parameters/JobID"
      responses:
        "200":
          description: Job progress snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobProgress"
        "404":
          $ref: "#/components/responses/NotFound"

  /jobs/{id}/midi:
    get:
      operationId: downloadMidi
      summary: Download the MIDI output for a completed job
      parameters:
        - $ref: "#/components/parameters/JobID"
      responses:
        "200":
          description: MIDI file
          content:
            audio/midi:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Job not yet completed
          $ref: "#/components/responses/Conflict"

  /health:
    get:
      operationId: healthCheck
      summary: Engine and dependency health check
      responses:
        "200":
          description: All healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
        "503":
          description: One or more dependencies unhealthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

components:
  parameters:
    JobID:
      name: id
      in: path
      required: true
      schema:
        type: string

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Operation not allowed in current state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    JobState:
      type: string
      enum: [queued, running, retrying, completed, failed, cancelled]

    Job:
      type: object
      required: [id, state, engine, device, attempt, max_attempts, created_at, updated_at]
      properties:
        id:
          type: string
        state:
          $ref: "#/components/schemas/JobState"
        youtube_url:
          type: string
          format: uri
        engine:
          type: string
        device:
          type: string
          enum: [cpu, cuda]
        start_sec:
          type: integer
          minimum: 0
        error_code:
          type: string
        error_message:
          type: string
        attempt:
          type: integer
        max_attempts:
          type: integer
        download_ms:
          type: integer
          format: int64
        inference_ms:
          type: integer
          format: int64
        total_ms:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time

    JobProgress:
      type: object
      required:
        - job_id
        - state
        - stage
        - progress_pct
        - attempt
        - max_attempts
        - cancelable
        - updated_at
      properties:
        job_id:
          type: string
        state:
          $ref: "#/components/schemas/JobState"
        stage:
          type: string
          description: Coarse execution stage for UI polling.
          enum: [queued, retrying, downloading, transcribing, finalizing, completed, failed, cancelled]
        progress_pct:
          type: integer
          minimum: 0
          maximum: 100
        attempt:
          type: integer
        max_attempts:
          type: integer
        cancelable:
          type: boolean
        error_code:
          type: string
        error_message:
          type: string
        download_ms:
          type: integer
          format: int64
        inference_ms:
          type: integer
          format: int64
        total_ms:
          type: integer
          format: int64
        updated_at:
          type: string
          format: date-time

    CreateJobRequest:
      type: object
      properties:
        youtube_url:
          type: string
          format: uri
        device:
          type: string
          enum: [auto, cpu, cuda]
          default: auto
        start_sec:
          type: integer
          minimum: 0
          default: 0
      anyOf:
        - required: [youtube_url]

    HealthResponse:
      type: object
      required: [status, checks]
      properties:
        status:
          type: string
          enum: [ok, degraded]
        checks:
          type: object
          additionalProperties:
            type: string

    ErrorResponse:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          type: string
        message:
          type: string